+++ eslint.config.js
@@ -3,7 +3,7 @@ import globals from 'globals';
 
 export default [
   js.configs.recommended,
-  { ignores: ['dist/**', 'lib/**', 'dist-zip/**', '**/*.min.js', 'assets/**', 'scripts/**', 'tests/**', 'vite.config.mjs'] },
+  { ignores: ['dist/**', 'lib/**', 'dist-zip/**', '**/*.min.js', 'assets/**', 'scripts/**', 'tests/**', 'vite.config.mjs', 'coverage/**'] },
   {
     languageOptions: {
       ecmaVersion: 2022,